apps/django-server/pyproject.toml
[tool.poetry]
name = "django-server"
version = "0.1.0"
description = "Simple Django application using Django Rest Framework"
authors = [
    "Artem Brytan <artem.brytan@dbbsoftware.com>",
    "Hryhoriy Bilenko <hryhoriy.bilenko@dbbsoftware.com>"
]
readme = "README.md"

[tool.ruff]
extend = "../../packages/python-configs/common/pyproject.toml"
exclude = ["**/migrations/*.py"]

[tool.poetry.dependencies]
python = "^3.11"
common = { path = "../../packages/python-configs/common" }
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^22.0.0"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
psycopg2 = "^2.9.9"
whitenoise = "^6.7.0"
django-health-check = "^3.18.3"

[tool.coverage.run]
branch = true
relative_files = true
omit = ["*/migrations/*", "*/tests/*", "manage.py", "core/settings/*", "*/__init__.py"]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "def __str__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
